Add files via upload
The purpose of this function is to utilize the bisection method in order to find the root of a given equation . Oftentimes, real life questions are difficult and hard to solve which is why root finding functions come in handy. In this specific instance, this method is utilized to find theta (in degrees) when there is a force of 150 N acting on a box with mass 25 kg. In order to use this function, one would enter the equation that needs to be solved and all the given variables that the function calls for. Upon entering these, the root, in this case theta, will be displayed on the screen.